Fix dashboard build to accomodate upstream changes
Commit 51bb0bb13f in upstream causes builds to fail on any webpack
error. To work around a typescript issue, commit 2181ed1340 adds
'index.d.ts' to the dashboard base directory. Since this is not included
in the packaged maven artifact, this commit adds the same file to the
rh-che dashboard directory and updates the build to copy it along with
the rest of the sources.

Should be undone when typescript is updated >3.5

Signed-off-by: Angel Misevski <amisevsk@redhat.com>